lv/lv 可以编译
git-svn-id: svn://23.27.117.66/h2/trunk@209 54bad739-14b1-1b4c-9595-0bb2987a2747

+++ lv/lv/Stream/OBufferStream.hpp
@@ -57,7 +57,7 @@ namespace lv
 
 		OBufferStream &	write(char const * data, streamsize size)
 		{
-			buffer::append(buf_, data, size);
+			buffer::append(buf_, data, static_cast<size_t>(size));
 
 			return reset_state();
 		}
